+++ kiplot/misc.py
@@ -21,6 +21,7 @@ NO_PCBNEW_MODULE = 16
 CORRUPTED_PCB = 17
 KICAD2STEP_ERR = 18
 WONT_OVERWRITE = 19
+PCBDRAW_ERR = 20
 
 CMD_EESCHEMA_DO = 'eeschema_do'
 URL_EESCHEMA_DO = 'https://github.com/INTI-CMNB/kicad-automation-scripts'
@@ -33,5 +34,6 @@ URL_KIBOM = 'https://github.com/INTI-CMNB/KiBoM'
 CMD_IBOM = 'generate_interactive_bom.py'
 URL_IBOM = 'https://github.com/INTI-CMNB/InteractiveHtmlBom'
 KICAD2STEP = 'kicad2step'
+PCBDRAW = 'pcbdraw'
 EXAMPLE_CFG = 'example.kiplot.yaml'
 AUTO_SCALE = 0