+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - The layers entry is much more flexible now.
   Many changes, read the README.md
+- PcbDraw output.
 - -e/--schematic option to specify any schematic (not just derived from the PCB
   name.
 - -x/--example option to generate a complete configuration example.